# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/message": {
            "post": {
                "description": "Create a message with expiration and optional password protection. Returns a Public ID and a Private Key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "Create a new one-time message",
                "parameters": [
                    {
                        "description": "Message Creation Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.CreateMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/message/fetch": {
            "post": {
                "description": "Retrieve and decrypt a message using its Public ID and Private Key. The message is deleted after retrieval.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "Retrieve a one-time message",
                "parameters": [
                    {
                        "description": "Message Retrieval Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.GetMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.CreateMessageRequest": {
            "type": "object",
            "required": [
                "expiration",
                "msg"
            ],
            "properties": {
                "expiration": {
                    "description": "Min 1 minute",
                    "type": "integer",
                    "minimum": 60
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "api.CreateMessageResponse": {
            "type": "object",
            "properties": {
                "private_key": {
                    "type": "string"
                },
                "public_id": {
                    "type": "string"
                }
            }
        },
        "api.GetMessageRequest": {
            "type": "object",
            "required": [
                "private_key",
                "public_id"
            ],
            "properties": {
                "private_key": {
                    "type": "string"
                },
                "public_id": {
                    "type": "string"
                }
            }
        },
        "api.GetMessageResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}